# openapi_spec.yaml
components:
  schemas:
    ActionItem:
      properties:
        action:
          title: Action
          type: string
        owner:
          title: Owner
          type: string
        status:
          default: pending
          title: Status
          type: string
        target_date:
          default: ''
          title: Target Date
          type: string
      required:
      - action
      - owner
      title: ActionItem
      type: object
...
    CorrectiveActionNotice:
      properties:
        can_id:
          default: ''
          title: Can Id
          type: string
        closed_at:
          default: ''
          title: Closed At
          type: string
        department:
          default: ''
          title: Department
          type: string
        description:
          title: Description
          type: string
        due_date:
          default: ''
          title: Due Date
          type: string
        hazard_id:
          default: ''
          title: Hazard Id
          type: string
        id:
          default: ''
          title: Id
          type: string
        issued_at:
          title: Issued At
          type: string
        issued_by:
          default: ''
          title: Issued By
          type: string
        priority:
          default: routine
          title: Priority
          type: string
        status:
          default: open
          title: Status
          type: string
        tenant_id:
          default: ''
          title: Tenant Id
          type: string
      required:
      - description
      title: CorrectiveActionNotice
      type: object
    CorrectiveActionPlan:
      properties:
        action_items:
          default: []
          items:
            $ref: '#/components/schemas/ActionItem'
          title: Action Items
          type: array
        approved_at:
          default: ''
          title: Approved At
          type: string
        can_id:
          default: ''
          title: Can Id
          type: string
        description:
          default: ''
          title: Description
          type: string
        id:
          default: ''
          title: Id
          type: string
        resource_requirements:
          default: ''
          title: Resource Requirements
          type: string
        review_notes:
          default: ''
          title: Review Notes
          type: string
        reviewed_by:
          default: ''
          title: Reviewed By
          type: string
        status:
          default: submitted
          title: Status
          type: string
        submitted_at:
          title: Submitted At
          type: string
        submitted_by:
          default: ''
          title: Submitted By
          type: string
      title: CorrectiveActionPlan
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  description: "Multi-tenant SMS aligned with ICAO Annex 19 / Doc 9859\n\nOperators:\n  - Sita Air (tenant_id: sita-air, API Key: sita-air-key-2026)\n\nRegulators:\n  - CAAN — Civil Aviation Authority of Nepal (tenant_id: caan, API Key: caan-reg-key-2026)"
  title: AviaSafe — Safety Management System
  version: 3.0.0
openapi: 3.1.0
paths:
...
  /api/v1/operator/{tenant_id}/cans:
    get:
      operationId: list_cans_api_v1_operator__tenant_id__cans_get
      parameters:
      - in: path
        name: tenant_id
        required: true
        schema:
          title: Tenant Id
          type: string
      - in: query
        name: status
        required: false
        schema:
          default: ''
          title: Status
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Cans
  /api/v1/operator/{tenant_id}/cans/{can_id}/cap:
    get:
      operationId: get_cap_api_v1_operator__tenant_id__cans__can_id__cap_get
      parameters:
      - in: path
        name: tenant_id
        required: true
        schema:
          title: Tenant Id
          type: string
      - in: path
        name: can_id
        required: true
        schema:
          title: Can Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Cap
    post:
      operationId: submit_cap_api_v1_operator__tenant_id__cans__can_id__cap_post
      parameters:
      - in: path
        name: tenant_id
        required: true
        schema:
          title: Tenant Id
          type: string
      - in: path
        name: can_id
        required: true
        schema:
          title: Can Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CorrectiveActionPlan'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Submit Cap
  /api/v1/operator/{tenant_id}/cans/{can_id}/cap/review:
    patch:
      operationId: review_cap_api_v1_operator__tenant_id__cans__can_id__cap_review_patch
      parameters:
      - in: path
        name: tenant_id
        required: true
        schema:
          title: Tenant Id
          type: string
      - in: path
        name: can_id
        required: true
        schema:
          title: Can Id
          type: string
      - in: query
        name: status
        required: true
        schema:
          title: Status
          type: string
      - in: query
        name: notes
        required: false
        schema:
          default: ''
          title: Notes
          type: string
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Review Cap
...
  /api/v1/operator/{tenant_id}/hazards/{hazard_id}/can:
    post:
      operationId: issue_can_api_v1_operator__tenant_id__hazards__hazard_id__can_post
      parameters:
      - in: path
        name: tenant_id
        required: true
        schema:
          title: Tenant Id
          type: string
      - in: path
        name: hazard_id
        required: true
        schema:
          title: Hazard Id
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CorrectiveActionNotice'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Issue Can
...